style(design change): redesigned the color scheme for both light/dark
I also changed the design of the calculator, and I will be
continuing to replace the button animation style to a more
simplistic design.

Instead of the button animating down, the filter changing, and
the background color changing, the buttons will simply have a
background color that glides across the button from one end
to the other on both hover and key press.